+++ packages/react-ui/src/theme/sections/components/forms.js
@@ -40,6 +40,41 @@ const checkboxRadioOverrides = {
   })
 };
 
+const LabelOverrides = {
+  root: {
+    position: 'static',
+    transform: 'none',
+    marginBottom: getSpacing(1),
+    ...themeTypography.caption,
+    fontWeight: 500,
+    color: commonPalette.text.primary
+  },
+  sizeSmall: {
+    marginBottom: getSpacing(0.5)
+  },
+  standard: {
+    marginBottom: 0
+  },
+  // Temporal workaroud to give a faster solution for a Workflows demand: https://app.shortcut.com/cartoteam/story/294539/distinction-among-parameters-optional-inputs-units-in-the-ui
+  // These styles avoid (by now) the need to review current forms and change those with a required mark.
+  // TODO: remored asterisk completely and use instead LabelWithIndicator component to mark also required ones.
+  asterisk: {
+    '&, &.Mui-error': {
+      color: 'transparent'
+    },
+    '&::after': {
+      content: '"(required)"',
+      marginLeft: getSpacing(-0.5),
+      color: commonPalette.text.secondary,
+      fontWeight: 400,
+
+      '.Mui-disabled &': {
+        color: commonPalette.text.disabled
+      }
+    }
+  }
+};
+
 export const formsOverrides = {
   // Checkbox
   MuiCheckbox: {
@@ -379,20 +414,12 @@ export const formsOverrides = {
   // Label
   MuiInputLabel: {
     styleOverrides: {
-      root: {
-        position: 'static',
-        transform: 'none',
-        marginBottom: getSpacing(1),
-        ...themeTypography.caption,
-        fontWeight: 500,
-        color: commonPalette.text.primary
-      },
-      sizeSmall: {
-        marginBottom: getSpacing(0.5)
-      },
-      standard: {
-        marginBottom: 0
-      }
+      ...LabelOverrides
+    }
+  },
+  MuiFormLabel: {
+    styleOverrides: {
+      ...LabelOverrides
     }
   },